Start refactoring to be a standalone package
Beginning to organize the repo more as a standalone package rather than a script. Running cd tests; spark-submit --master local[2] --packages com.databricks:spark-csv_2.11:1.3.0 test_psaml.py; cd ..;
DATA:

DataFrame[C0: double, C1: double, C2: double, C3: double, C4: string]
+----------+-----+-----------------+
|prediction|label|         features|
+----------+-----+-----------------+
|       0.0|  0.0|[4.3,2.0,1.0,0.1]|
|       0.0|  0.0|[4.9,2.0,1.0,0.1]|
|       0.0|  0.0|[5.5,2.0,1.0,0.1]|
|       0.0|  0.0|[6.1,2.0,1.0,0.1]|
|       0.0|  0.0|[6.7,2.0,1.0,0.1]|
|       0.0|  0.0|[7.3,2.0,1.0,0.1]|
|       0.0|  0.0|[7.9,2.0,1.0,0.1]|
|       1.0|  0.0|[4.3,2.4,2.0,0.5]|
|       1.0|  0.0|[4.9,2.4,2.0,0.5]|
|       1.0|  0.0|[5.5,2.4,2.0,0.5]|
|       1.0|  0.0|[6.1,2.4,2.0,0.5]|
|       1.0|  0.0|[6.7,2.4,2.0,0.5]|
|       1.0|  0.0|[7.3,2.4,2.0,0.5]|
|       1.0|  0.0|[7.9,2.4,2.0,0.5]|
|       1.0|  0.0|[4.3,2.8,3.0,0.9]|
|       1.0|  0.0|[4.9,2.8,3.0,0.9]|
|       1.0|  0.0|[5.5,2.8,3.0,0.9]|
|       1.0|  0.0|[6.1,2.8,3.0,0.9]|
|       1.0|  0.0|[6.7,2.8,3.0,0.9]|
|       1.0|  0.0|[7.3,2.8,3.0,0.9]|
+----------+-----+-----------------+
only showing top 20 rows

Root Mean Squared Error (RMSE) on test data = 1.45336
VectorIndexer_4346bd34b59ec5f9c291 from within the main directory will run the iris test, and we can use that to define a single test for the time being. The idea I have is to kind of use it as a test-driven development goal, and then we refactor psaml to be a standalone package that can then be called by the test. This involves isolating the analysis part of the  file, and then removing that bit from the test case. The package will expect a model and data, which are generated and passed in via the test script, and then the package will analyze and return the sensitivity analysis. Needless to say, most of that has yet to be done, but this is a start.