This is what the last commit was supposed to say:
Beginning to organize the repo more as a standalone package rather than a script. Running `make test` from within the main directory will run the iris test, and we can use that to define a single test for the time being. The idea I have is to kind of use it as a test-driven development goal, and then we refactor psaml to be a standalone package that can then be called by the test. This involves isolating the analysis part of the `iris_analysis.py` file, and then removing that bit from the test case. The package will expect a model and data, which are generated and passed in via the test script, and then the package will analyze and return the sensitivity analysis. Needless to say, most of that has yet to be done, but this is a start."